// src/app/core/interfaces/pessoa.interface.ts
export interface PessoaInterface {
    nome: string;
    cpf: string;
    tipo_documento: number;
    documento: string;
    orgao_emissor: string;
    uf_orgao_emissor: string;
    data_nascimento: string;
    is_brasileiro: boolean;
    nacionalidade: number;
    naturalidade_estrangeira: string;
    uf_naturalidade: number;
    naturalidade: number;
    estado_civil: number;
    nome_mae: string;
    nome_pai: string;
    sexo: string;
}